ai/requirements.txt
python-dotenv==1.0.0
google-api-python-client==2.118.0
requests==2.31.0
